Split methods.go into per-method files.
